Separated seals
