+++ index.html
@@ -7,6 +7,7 @@
    </script>
    <!-- Global site tag (gtag.js) - Google Analytics -->
    <script async src="https://www.googletagmanager.com/gtag/js?id=UA-110108110-1"></script>
+   <script type="text/javascript" src="thirdparty/lodash.js"></script>
    <script>
      window.dataLayer = window.dataLayer || [];
      function gtag(){dataLayer.push(arguments);}
@@ -54,7 +55,7 @@
    }
    </style>
 <script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
-<script type="text/javascript" src="main.js"></script>
+<script type="module" src="main.js"></script>
 <script type="text/javascript" src="Sentimental/wordLists/afinn.js"></script>
 <script type="text/javascript" src="Sentimental/sentimental.js"></script>
 </head>
