+++ misc.js
@@ -1,0 +1,37 @@
+var _units = {
+  "градус" : ["градуса", "градусов"],
+  "метр" : ["метра", "метров"],
+  "миллиметр" : ["миллиметра", "миллиметров"],
+  "процент": ["процента", "процентов"],
+  "тюлень": ["тюленя", "тюленей"],
+};
+
+for(var u in _units)
+  if(_units.hasOwnProperty(u))
+     _units[u].unshift(u);
+
+export function declinateUnit(value, unit){
+  var a = _units[unit];
+  if(value < 0)
+    value = -value;
+  var lastdigit = value % 10;
+  var lasttwodigits = value % 100;
+  if(lasttwodigits >= 10 && lasttwodigits <= 20)
+    lastdigit = 5;
+
+  switch(lastdigit){
+    case 0:
+    case 5:
+    case 6:
+    case 7:
+    case 8:
+    case 9:
+      return a[2];
+    case 1:
+      return a[0];
+    case 2:
+    case 3:
+    case 4:
+      return a[1];
+  }
+}
